+++ sdk/dasPayment/log/index.js
@@ -1,0 +1,4 @@
+const log = require('./log.js');
+
+exports.get = log.get;
+exports.query = log.query;