+++ tests/testDasPaymentLog.js
@@ -1,0 +1,31 @@
+const assert = require('assert');
+const starkbank = require('../index.js');
+
+starkbank.user = require('./utils/user').exampleProject;
+
+
+describe('TestDasPaymentLogGet', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let i = 0;
+        const logs = await starkbank.dasPayment.log.query({limit: 5});
+        for await (let log of logs) {
+            assert(typeof log.id == 'string');
+            i += 1;
+        }
+        assert(i === 5);
+    });
+});
+
+
+describe('TestDasPaymentLogInfoGet', function(){
+    this.timeout(10000);
+    it('test_success', async () => {
+        let logs = await starkbank.dasPayment.log.query({limit: 1});
+        for await (let log of logs) {
+            assert(typeof log.id == 'string');
+            log = await starkbank.dasPayment.log.get(log.id);
+            assert(typeof log.id == 'string');
+        }
+    });
+});